+++ api/src/crons/reminderWaitingCorrection.js
@@ -1,3 +1,4 @@
+const { differenceInDays, startOfDay } = require("date-fns");
 const { capture } = require("../sentry");
 const YoungModel = require("../models/young");
 const { sendTemplate } = require("../sendinblue");
@@ -7,25 +8,10 @@ const { SENDINBLUE_TEMPLATES, YOUNG_STATUS } = require("snu-lib");
 exports.handler = async () => {
   try {
     let countNotice = 0;
-    const now = new Date();
-    now.setUTCHours(0, 0, 0, 0);
-
-    const threeDaysAgo = new Date(now);
-    threeDaysAgo.setDate(now.getDate() - 3);
-
-    const sevenDaysAgo = new Date(now);
-    sevenDaysAgo.setDate(now.getDate() - 7);
+    const today = startOfDay(new Date());
 
     const cursor = await YoungModel.find({
-      correctionRequests: {
-        $elemMatch: {
-          status: "SENT",
-          sentAt: {
-            $gte: sevenDaysAgo,
-            $lte: threeDaysAgo,
-          },
-        },
-      },
+      "correctionRequests.status": "SENT",
       status: YOUNG_STATUS.WAITING_CORRECTION,
     }).cursor();
 
@@ -34,9 +20,10 @@ exports.handler = async () => {
     await cursor.eachAsync(async function (young) {
       try {
         for (const request of young.correctionRequests) {
-          const sentAt = new Date(request.sentAt);
-          // On compare les jours sans les heures pour faire au plus juste
-          if (sentAt.toISOString().split("T")[0] === threeDaysAgo.toISOString().split("T")[0] || sentAt.toISOString().split("T")[0] === sevenDaysAgo.toISOString().split("T")[0]) {
+          const sentAt = startOfDay(new Date(request.sentAt));
+          const daysDiff = differenceInDays(today, sentAt);
+
+          if (daysDiff === 3 || daysDiff === 7) {
             countNotice++;
             await sendTemplate(SENDINBLUE_TEMPLATES.young.PHASE_1_WAITING_CORRECTION_REMINDER, {
               emailTo: [{ name: `${young.firstName} ${young.lastName}`, email: young.email }],
